--- Desktop_Gremlin/App.xaml
@@ -1,9 +1,0 @@
-﻿<Application x:Class="Desktop_Gremlin.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:Desktop_Gremlin"
-             StartupUri="SpriteSheet.xaml">
-    <Application.Resources>
-         
-    </Application.Resources>
-</Application>
