+++ Desktop_Gremlin/Desktop_Gremlin.sln
@@ -1,0 +1,16 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Desktop_Gremlin", "Desktop_Gremlin.csproj", "{3C1524A2-B440-45A1-8AA7-EB59337CDA87}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3C1524A2-B440-45A1-8AA7-EB59337CDA87}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3C1524A2-B440-45A1-8AA7-EB59337CDA87}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3C1524A2-B440-45A1-8AA7-EB59337CDA87}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3C1524A2-B440-45A1-8AA7-EB59337CDA87}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
