--- Desktop_Gremlin/SpriteSheet.xaml
@@ -1,49 +1,0 @@
-﻿<Window x:Class="Desktop_Gremlin.SpriteSheet"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Desktop_Gremlin"
-        mc:Ignorable="d"
-        Title="MainWindow" Height="235" Width="210"
-        WindowStyle="None"
-        AllowsTransparency="True"
-        Background="Transparent"
-        Topmost="True"
-        ResizeMode="NoResize"  BorderThickness="0" BorderBrush="Black" KeyDown="Window_KeyDown" KeyUp="Window_KeyUp">
-    <Grid MouseLeftButtonDown="Canvas_MouseLeftButtonDown" Margin="0,0,10,10" >
-        <Image Name="SpriteImage"
-           Width="160" Height="162"
-           HorizontalAlignment="Left"
-           VerticalAlignment="Top"
-           Canvas.Top="20"  Margin="15,35,0,0"/>
-        <Label x:Name="SpriteLabel"
-           Content=""
-           Background="Transparent"
-           Foreground="Transparent"
-           FontWeight="Bold"
-           FontSize="14"
-           VerticalContentAlignment="Center" RenderTransformOrigin="-0.145,-0.349" HorizontalAlignment="Left" VerticalAlignment="Top" Width="135" Canvas.Left="45" Canvas.Top="6" Margin="115,1,0,0"/>
-        <Border x:Name="SpeechBubbleBorder"
-                Background="Black"
-                BorderBrush="White"
-                BorderThickness="1.5"
-                Width="120" Height="74"
-                Opacity="0"
-                Padding="5"
-                HorizontalAlignment="Left"
-                VerticalAlignment="Top"
-                Visibility="Visible"
-                RenderTransformOrigin="0,0" Margin="75,6,0,0"
-    >
-            <TextBlock x:Name="SpriteSpeech"
-                   Text=""
-                   Foreground="White"
-                   FontWeight="Bold"
-                   FontSize="15"
-                   FontFamily="Agency FB"
-                   TextWrapping="Wrap"/>
-        </Border>
-
-    </Grid>
-</Window>
