+++ Desktop_Gremlin/Sprites/SpriteSheet/emote1.png
