+++ Desktop_Gremlin/Sprites/SpriteSheet/emote2.png
