+++ Desktop_Gremlin/Sprites/SpriteSheet/run_left.png
