+++ Desktop_Gremlin/config.txt
@@ -9,4 +9,4 @@ WALK_RIGHT_FRAME=18
 EMOTE1_FRAME_COUNT=61
 EMOTE2_FRAME_COUNT=23               
 EMOTE3_FRAME_COUNT=95
-FRAME_RATE=30
+FRAME_RATE=50
